refactor: move Customer to services (#141)
* refactor: move Customer to services

* refactor: add missed file
